moved to emacs editing and building through build.bat
this is just so much faster on my machine so might as well stick to it.
