+++ code/build.bat
@@ -1,0 +1,5 @@
+@Echo off
+IF NOT EXIST ..\build mkdir ..\build
+pushd ..\build
+cl -DHANDMADE_WIN32=1 -FC -Zi ..\code\win32_handmade.cpp user32.lib gdi32.lib
+popd
