+++ misc/emacs.bat
@@ -1,0 +1,2 @@
+@Echo off
+w:\emacs\bin\runemacs.exe -q -l w:\hmh\misc\.emacs
