+++ misc/quit.bat
@@ -1,0 +1,4 @@
+@Echo off
+c:
+subst w: /D
+call exit
