+++ misc/shell.bat
@@ -1,0 +1,5 @@
+@Echo off
+subst w: C:\Projects
+call "C:\Program Files (x86)\Microsoft Visual Studio\2017\Community\VC\Auxiliary\Build\vcvarsall.bat" x64
+set PATH=w:\hmh\misc;%path%
+w:
