--- win32_handmade/packages.config
@@ -1,6 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<packages>
-  <package id="Doxygen" version="1.8.13" targetFramework="native" />
-  <package id="Microsoft.CppCoreCheck" version="14.0.24210.1" targetFramework="native" />
-  <package id="Microsoft.Gsl" version="0.1.2.1" targetFramework="native" />
-</packages>
